Copy TraceComponentImpl into Android, Java 7, and Java 8 directories.
This change will allow TraceComponentImpl to have different implementations for
the different Java versions.
